moved hero from template to block
